backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Document processing
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==10.2.0
python-docx==1.1.0

# AI/LLM
openai==1.12.0
litellm==1.27.0
langchain==0.1.5
tiktoken==0.5.2

# Data processing
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
PyYAML==6.0.1

# Utilities
aiofiles==23.2.1
httpx==0.26.0
tenacity==8.2.3
